% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{print.projmap_compare}
\alias{print.projmap_compare}
\title{Custom Print Method for compare_files}
\usage{
\method{print}{projmap_compare}(x, ...)
}
\arguments{
\item{x}{An object of class projmap_compare}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Custom Print Method for compare_files
}
